requirements.txt
streamlit
qianfan
paddlepaddle
paddleocr
pymilvus
pandas
python-dotenv
numpy
opencv-python